+++ src/Services/PaymentService.php
@@ -5,7 +5,6 @@ use Ordercloud\Requests\Exceptions\OrdercloudRequestException;
 use Ordercloud\Requests\Payments\CreateCashOnDeliveryPaymentRequest;
 use Ordercloud\Requests\Payments\CreateCreditCardPaymentRequest;
 use Ordercloud\Requests\Payments\Entities\CreditCard;
-use Ordercloud\Requests\Payments\Exceptions\CreditCardPaymentFailedException;
 use Ordercloud\Requests\Payments\GetPaymentThreeDSecureRequest;
 
 class PaymentService extends OrdercloudService
@@ -28,11 +27,12 @@ class PaymentService extends OrdercloudService
      * @param int        $orderID
      * @param string     $budgetPeriod
      * @param bool       $threeDSecure
+     * @param string     $threeDSecureReturnUrl
      */
-    public function createCreditCardPayment($paymentGateway, CreditCard $card, $orderID, $budgetPeriod = '0', $threeDSecure = false)
+    public function createCreditCardPayment($paymentGateway, CreditCard $card, $orderID, $budgetPeriod = '0', $threeDSecure = false, $threeDSecureReturnUrl = null)
     {
         $this->request(
-            new CreateCreditCardPaymentRequest($paymentGateway, $card, $orderID, $budgetPeriod, $threeDSecure)
+            new CreateCreditCardPaymentRequest($paymentGateway, $card, $orderID, $budgetPeriod, $threeDSecure, $threeDSecureReturnUrl)
         );
     }